refactor: hover 시 색 변경을 위해 svg 동적 생성
