+++ components/ColumnItem/ui/createItemButtonContainer.js
@@ -1,29 +1,10 @@
 import { IMAGE } from "../../../assets/index.js";
-import { createButton, createElement, createImg } from "../../../dom.js";
-
-const createSvg = ({ className, ...props }) => {
-  const SVG_NS = "http://www.w3.org/2000/svg";
-
-  const $svg = document.createElementNS(SVG_NS, "svg");
-  $svg.setAttribute("xmlns", SVG_NS);
-  $svg.setAttribute("class", className);
-
-  props &&
-    Object.entries(props).forEach(([key, value]) => {
-      $svg.setAttribute(key, value);
-    });
-
-  const path = document.createElementNS(SVG_NS, "path");
-  path.setAttribute(
-    "d",
-    "M7.2 18L6 16.8L10.8 12L6 7.2L7.2 6L12 10.8L16.8 6L18 7.2L13.2 12L18 16.8L16.8 18L12 13.2L7.2 18Z"
-  );
-  path.setAttribute("fill", "#A0A3BD");
-
-  $svg.appendChild(path);
-
-  return $svg;
-};
+import {
+  createButton,
+  createDeleteSvg,
+  createElement,
+  createImg,
+} from "../../../dom.js";
 
 const createItemButtonContainer = () => {
   const $itemButtonContainer = createElement("div", {
@@ -34,7 +15,7 @@ const createItemButtonContainer = () => {
     handleClick: () => alert("Delete button clicked"),
   });
 
-  const $deleteImg = createSvg({
+  const $deleteImg = createDeleteSvg({
     className: "delete__img",
     width: "24",
     height: "24",
